generalize http tests
